Relax virtual address space in section padding
What if VirtualSize > SizeOfRawData?
This means the remainder of the section should be zero filled,
it is not possible to get a reference to the zero filled data.

What if SizeOfRawData > VirtualSize?
This is usually because SizeOfRawData is aligned to file alignment
while VirtualSize is exact to the last used byte.
Logically they are part of the virtual address space but technically OOB.

This patch changes the previous behavior and allows the address space
between VirtualSize and SizeOfRawData to be referenced.